+++ tests/testthat/test-translate-math.R
@@ -31,11 +31,11 @@ test_that("binary minus subtracts", {
 # log ---------------------------------------------------------------------
 
 test_that("log base comes first", {
-  expect_equal(translate_sql(log(x, 10)), sql('log(10.0, "x")'))
+  expect_equal(translate_sql(log(x, 10)), sql('LOG(10.0, "x")'))
 })
 
 test_that("log becomes ln", {
-  expect_equal(translate_sql(log(x)), sql('ln("x")'))
+  expect_equal(translate_sql(log(x)), sql('LN("x")'))
 })
 
 test_that("sqlite mimics two argument log", {
@@ -43,8 +43,8 @@ test_that("sqlite mimics two argument log", {
     translate_sql(..., con = src_memdb()$con)
   }
 
-  expect_equal(translate_sqlite(log(x)), sql('log(`x`)'))
-  expect_equal(translate_sqlite(log(x, 10)), sql('log(`x`) / log(10.0)'))
+  expect_equal(translate_sqlite(log(x)), sql('LOG(`x`)'))
+  expect_equal(translate_sqlite(log(x, 10)), sql('LOG(`x`) / LOG(10.0)'))
 })
 
 test_that("postgres mimics two argument log", {
@@ -53,6 +53,6 @@ test_that("postgres mimics two argument log", {
   }
 
   expect_equal(translate_postgres(log(x)), sql('ln(\"x\")'))
-  expect_equal(translate_postgres(log(x, 10)), sql('log(\"x\") / log(10.0)'))
-  expect_equal(translate_postgres(log(x, 10L)), sql('log(\"x\") / log(10)'))
+  expect_equal(translate_postgres(log(x, 10)), sql('LOG(\"x\") / LOG(10.0)'))
+  expect_equal(translate_postgres(log(x, 10L)), sql('LOG(\"x\") / LOG(10)'))
 })